refactor: move mock files to a new mock folder at root
