+++ __mocks__/authentication.mock.ts
@@ -1,7 +1,17 @@
-import faker from '@faker-js/faker';
-import { mockAccountModel } from '~/domain/mocks';
+import { faker } from '@faker-js/faker';
 import { AccountModel } from '~/domain/models';
 import { Authentication, AuthenticationParams } from '~/domain/usecases';
+
+export const mockAuthentication = (): AuthenticationParams => ({
+  email: faker.internet.email(),
+  password: faker.internet.password(),
+});
+
+export const mockAccountModel = (): AccountModel => ({
+  accessToken: faker.datatype.uuid(),
+});
+
+// presentation layer
 
 export class AuthenticationSpy implements Authentication {
   params: AuthenticationParams;
