+++ __mocks__/field-validation.mock.ts
@@ -3,7 +3,7 @@ import { FieldValidation } from '~/validation/protocols/field-validation';
 export class FieldValidationSpy implements FieldValidation {
   error: Error;
   constructor(readonly field: string) {}
-  validate(value: string): Error {
+  validate(_value: string): Error {
     return this.error;
   }
 }
