+++ __mocks__/http-post.mock.ts
@@ -1,6 +1,6 @@
 import faker from '@faker-js/faker';
-import { makeRandomObject } from '~/infra/mocks';
 import { HttpPostParams } from '~/data/protocols/http';
+import { makeRandomObject } from './axios.mock';
 
 export const mockPostRequest = (): HttpPostParams<unknown> => ({
   url: faker.internet.url(),
