+++ __mocks__/use-authentication.mock.ts
@@ -1,3 +1,3 @@
-jest.mock('../hooks', () => ({
+jest.mock('../src/presentation/hooks', () => ({
   useAuthentication: jest.fn(() => [false, jest.fn()]),
 }));
