--- src/data/mocks/index.ts
@@ -1,2 +1,0 @@
-export * from './http-client.mock';
-export * from './http-post.mock';
