--- src/domain/mocks/authentication.mock.ts
@@ -1,12 +1,0 @@
-import { faker } from '@faker-js/faker';
-import { AccountModel } from '../models';
-import { AuthenticationParams } from '../usecases';
-
-export const mockAuthentication = (): AuthenticationParams => ({
-  email: faker.internet.email(),
-  password: faker.internet.password(),
-});
-
-export const mockAccountModel = (): AccountModel => ({
-  accessToken: faker.datatype.uuid(),
-});
