--- src/domain/mocks/index.ts
@@ -1,1 +1,0 @@
-export * from './authentication.mock';
