+++ src/infra/http/axios-http-client/index.ts
@@ -1,0 +1,1 @@
+export * from './axios-http-client';
