--- src/infra/mocks/index.ts
@@ -1,1 +1,0 @@
-export * from './axios.mock';
