--- src/presentation/mocks/index.ts
@@ -1,2 +1,0 @@
-export * from './validation.mock';
-export * from './authentication.mock';
