--- src/validation/mocks/index.ts
@@ -1,1 +1,0 @@
-export * from './field-validation.mock';
